+++ docs/release-notes.md
@@ -1,5 +1,24 @@
 # Release Notes
 
+## Unreleased
+
+### Added
+
+- Allow submitting paused tasks.
+- Allow changing in-place attributes of tasks when submitting with `add_task`.
+
+### Fixed
+
+- `remove_all_tasks` didn't check the store of tasks when pending_tasks was used (stopped scheduler).
+
+### Changed
+
+- Replace `UndefinedType` with `Undefined`. Shim provided for backwards compatibility.
+- `add_task` has now more arguments with undefined as default.
+- `pending_tasks` has now no more store alias in it.
+- `tzlocal` is now optional.
+- Tasks use the timezone of the scheduler for their triggers which require a timezone.
+
 ## 0.10.1
 
 ### Added